Showing the search term (see description for Jira description)
The search term is not retained in the results page. So it is hard to tell what search term are the search results related to. Is it possible to display the search term in the search results page like the results page in some of our other applications
For Ex: Section508.gov(fixed)